Add a python notebook to calculate demand for each station
Actual demand for each station has yet to be set.

Follow the code to determine the demand category for each station, which can be: "Lowest" , "Low", "Medium", "Highest".

The plan is to set an arbitrary value for demand using %X*capacity_of_station,
where %X is determined by the demand category of that station.

E.g. if demand category "Lowest" sets demand factor to 0.10, then demand for a station with capacity 100 is 10.