add a N*N distance matrix in unit: US feet
Hi, I've previously recomputed the distance matrix as part of my attempt on stage 2. Just committing it for now.

The station_ids rows and column are not sorted. They are ordered as they were in the file manhattan_stations.csv.

The units and formula is in US feet using (EPSG:2263) and manhattan distance respectively.